pyproject.toml
[tool.poetry]
name = "ollama-rag"
version = "0.1.0"
description = "Simple RAG using Ollama as LLM engine"
authors = ["Alan Tan <alan@alantan.com>"]
license = "private"

readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
ollama = "^0.4.7"
nest-asyncio = "^1.6.0"
llama-index = "^0.12.14"
llama-index-llms-ollama = "^0.5.0"
llama-index-embeddings-ollama = "^0.5.0"
qdrant-client = "^1.13.1"
llama-index-graph-stores-neo4j = "^0.4.6"
llama-index-vector-stores-redis = "^0.4.0"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
neo4j = "^5.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
